add infer_mean for faster posterior mean computation
Summary:
Added an option to compute the mean over the posterior samples and return these means from C++ rather than sending back the full samples which was very expensive.

On the U-Topic benchmark sending out the means instead of the full samples gave a 2x improvement for 1000 samples, a 3x improvement, and for 10K samples a 4x improvement.

Reviewed By: nazanint

Differential Revision: D17204567

fbshipit-source-id: 7fdcf4631f45f8c263a233bca797084a8f15c8d9