add NETNORAD benchmark with boolean loss
Summary:
I have added a benchmark using real data from NET-NORAD and using a model of boolean observed loss variables. More explanations of the model are in the notebook n140582. The benchmark can be run with the command:

    buck run mode/opt //beanmachine/facebook/graph/netnorad:boolean-
    loss

OR for the C++ interface directly:

    buck run mode/opt //beanmachine/facebook/graph/netnorad:boolean-loss-cpp

Also, changed the default inference from Rejection to Gibbs. Otherwise, I was accidentally using rejection sampling and this was causing confusion.

Reviewed By: nazanint

Differential Revision: D17179747

fbshipit-source-id: fec9f55c1f20c894175c2859101a074cafbd21ac