Add helper method for rendering Python objects as trees when debugging.
Summary:
I've added a handy helper method that can take an arbitrary Python object and dump it as a tree.  Usage:

    > from beanmachine.ppl.utils.treeprinter import print_tree

    > print(print_tree({"foo": 2, "bar": {"blah": [2, 3, {"abc": (6, 7, (5, 5, 6))}]}}))

    dict
    ├─foo
    │ └─2
    └─bar
      └─blah
        ├─2
        ├─3
        └─dict
          └─abc
            ├─6
            ├─7
            └─tuple
              ├─5
              ├─5
              └─6

It automatically makes a tree structure for dictionaries, lists, tuples and named tuples; everything else is treated as a leaf. If that' doesn't suit, you can provide your own get-the-children and node-to-string function as well.

Note that the data structure passed in must be a tree; it will go into an infinite loop if given a cyclic graph.

Reviewed By: janewangfb, nazanint

Differential Revision: D17505387

fbshipit-source-id: 264e922e72d2a214bf7857d2909a3d8003ecc98d