Add a helper class for visualizing Python objects as graphs
Summary:
I've added a class that will help us visualize graphs of objects for debugging or demo purposes.

We have a mutable `DotBuilder` class which allows for imperatively adding nodes, edges and subgraphs; the result is then a string in the DOT language, which can be turned into a graphic with `dot`, or `webgraphviz.com` or `https://our.intern.facebook.com/intern/graphviz/`.

I've also added a helper function for simple cases; `print_graph` takes a list of roots and functions for producing the outgoing labeled edges, node identifiers and node labels, and produces a graph.

For example:

        bar = {"blah": [2, 3, {"abc": (6, 7, (5, 5, 6))}]}
        d = {"foo": 2, "bar1": bar, "bar2": bar}
        d["self"] = d
        print_graph([d])

produces the graph below:

{F210813866}

Reviewed By: janewangfb

Differential Revision: D17509283

fbshipit-source-id: 364e7f79320ba2ddea5c82b658c83ce23e1a9619