change iterations to trials
Summary:
In pplbench, we could run inference multiple times and we controlled this by a parameter called iterations. But this is confusing because iterations is also used to denote the number of times we sample from the posterior in our inference algorithm.

Hence, this diff changes the term iterations to trials to avoid confusion

Reviewed By: nimar

Differential Revision: D17818671

fbshipit-source-id: 92668dcbfbe8efdca444ea9f35a62100c4dbdab9