bug fixes in the robust regression model
Summary:
I found some bugs in the robust regression model - both in the definition, and in the pyro implementation of robust regression.

In the model definition, to evaluate posterior log likelihood, we need to sum over all observations, not average.

In the pyro implementation, the order in which model args were assigned values was incorrect.

Reviewed By: nimar

Differential Revision: D17812408

fbshipit-source-id: 6312a3a5f4e4534de34f391672a249b0861c12b5