update pyro implementation of BLR
Summary:
The main change here is I use the TracePosterior class to extract samples from the approximate posterior instead of the TracePredictive class.

When I looked through the SVI documentation more, I realized that TracePredictive "Generates and holds traces from the posterior predictive distribution" versus TracePosterior which "collects a bag of execution traces from the approximate posterior".

We probably don't want to be sampling from the posterior predictive distribution and want to sample from the approximate posterior instead.

I also added a plot of ELBO loss to test for convergence.

Reviewed By: nimar

Differential Revision: D17902595

fbshipit-source-id: 8a7e3c8d0b8c32ad00f89c42bcb4b300beaa5266