Implement Python helper functions for computing equivalence classes
Summary:
In an upcoming diff I will need to be able to partition a sequence into equivalence classes, so I've written code to solve the general problem.

There are two methods; one takes a sequence of values and an equivalence relation; an equivalence relation is a function that takes two values and returns a bool that says if the two values are "equivalent" or not. An equivalence relation is required to be reflexive -- all values are equivalent to themselves -- and symmetric -- if A is equivalent to B then B is equivalent to A -- and transitive -- A equivalent to B and B equivalent to C implies A equivalent to C. We can then take all the items in the sequence and put all the items that are equivalent to each other into an equivalence class, and then return the list of classes.

The other takes a sequence of values and an equivalence kernel; an equivalence kernel is a function that takes a value and returns a value that uniquely identifies which equivalence class it is in.

Reviewed By: nimar

Differential Revision: D17608727

fbshipit-source-id: cdce60ca6b668d8072ad12670dbbfd20ff205635