Implement unique-naming helper class in utils
Summary:
For my graph visualizer I needed a service that produces *consistent* unique names for a set of arbitrary objects -- they have to be consistent so that we can write test cases.

I'm going to need that service in some other code in an upcoming diff, so I've moved it to its own class.

Reviewed By: nimar

Differential Revision: D17635224

fbshipit-source-id: 8213d923ac1d268f334c466829795fbca7ac93a2