Random Walk MH with real number support
Summary:
Implementing initial RWMH class as a subclass of AncestralSingleSiteMH. Additionally, add the unit test cases, as well as two Normal-Normal conjugate tests.

Adding the first random variable support, those with support over the full real number line.

This is replacing an old diff, primarily because the codebase was refactored, and the conjugate-test class was modified.

Reviewed By: nimar

Differential Revision: D17796383

fbshipit-source-id: c6a5375661c744b6ac5cb9682eca99ca3c586900