Add yet another graph abstraction, this time for plate diagramming
Summary:
I've been exploring the question "can we automatically generate the plate diagram for a world?" or, slightly less ambitious, can we at least generate a more concise graphical representation of a world than simply building a graph with one node per random variable?

I believe we can, but doing so requires us to be able to (1) perform common-subgraph analysis on an existing graph, and (2) transform all instances of common subgraphs into "plates", while preserving the edges that both enter and leave the subgraph.

Unfortunately neither graph abstraction we have already -- the "world", and my "dot builder" for rendering graphs to DOT -- will be easy to make support these operations, and besides, we do not want to edit the world anyways.

Therefore I have made a *third* graph abstraction. We might wish in the future to make the "world" use this abstraction and have it generate DOT files directly, and therefore eliminate two somewhat redundant abstractions, but I will consider those possibilities in later diffs.

The abstraction I've created is `Graph[T]`, which represents a digraph where the nodes are arbitrary objects of type T. (T will be `RandomVariable` when we come to generating plates.)  A graph is a tree of plates. Every node in the graph belongs to exactly one plate.  Edges may arbitrarily span plates.

The API is fluent; adding a plate returns the plate; adding anything else returns the container so that another thing may be added.

If an edge is added and either node is not in the graph, it is added to the current container.

The test case given represents this diagram. (Note that this is a cyclic graph and therefore not a legal example of plate notation; the API is rich enough to represent arbitrary digraphs, but when we come to common subgraph elimination, I will at that point make the simplifying assumption that we're dealing with a DAG.)

{F212779355}

Reviewed By: nimar

Differential Revision: D17639612

fbshipit-source-id: 50ce863bb97ce017d5d45b98920dcc453005ec41