Start implementing graph isomorphism for plate diagram generation
Summary:
To generate a plate diagram from a DAG the first thing we'll need to do is find the common isomorphic sub-dags.

This diff begins working on that algorithm; here we can take two nodes in a graph and determine if the pair is isomorphic or not.

Now, I hear your objection: dag isomorphism is GI-complete and no polynomial-time algorithm is known, and yet I seem to have written one.

The algorithm I am using is not perfect; in the event of an MD5 hash collision, it will say that two non-isomorphic graphs are isomorphic.  I am willing to live with that possibility.

Reviewed By: nimar

Differential Revision: D17645198

fbshipit-source-id: 1d116f6f7cd0ba027d3553de1e16c73143632712