Implement initial merging of two isomorphic graph nodes.
Summary:
If we have two isomorphic nodes we can "merge" them by deleting one, and the incoming and outgoing edges of the deleted node become those of the surviving node. For example, merging the two isomorphic "2" nodes in this:

{F213519613}

turns it into this:

{F213516733}

If we then recurse on the two "5" children of the surviving node, we will de-duplicate the entire set of descendants; that will come in a later diff.

Note that figuring out where the "plates" are and moving the merged nodes into subgraphs is a harder problem which I will solve at a later time. Right now I'm just trying to get something that works more or less.

While testing I noticed two bugs in the graph code, which I've fixed.

Reviewed By: nimar

Differential Revision: D17675015

fbshipit-source-id: 245fc3a01a9240eedd71def0220164009b7785e9