Implement merging of all isomorphic children of a node
Summary:
We can now take a node in a graph, partition the children into collections of isomorphic subgraphs, and then merge all the isomorphic subgraphs.

For example, if we have:

{F213569178}

and we ask that all children of "2" be merged, then we merge the 3s and the 4s, but not the 5s:

{F213569910}

We'll extend this first to all descendants, and then to the entire graph, in future diffs.

Again, this is not yet sophisticated enough to do plate notation; I just want to first see if a straightforward graph simplification algorithm gives any kind of good results.

Reviewed By: nimar

Differential Revision: D17678324

fbshipit-source-id: 4d3bb1ab041da090d4abecd5b5a4572b1e93da25