RWMH with Beta noise for cts. finite intervals (#46)
Summary:
Pull Request resolved: https://github.com/facebookincubator/BeanMachine/pull/46

Use Beta-distributed noise for proposals on a finite interval of the real number line. This primarily requres rescaling the beta distribution between [0,1] and whichever necessary [a,b].

I set a floor for the expectation `mu` of the beta distribution at `1e-5`, although in theory this value can fall anywhere in (0,1). Setting this floor resolved some strange divide-by-zero errors which were not trigged in Python, but fell through to Caffe2/Aten.

Reviewed By: nazanint

Differential Revision: D17957624

fbshipit-source-id: 1a7ee7c2b58c88fba26ab1d24bb928a318da36b6