Buck TARGETS, and first test, for PPLBench+BeanMachine
Summary:
A Buck test which runs the PPLBench module for inference by Bean Machine on some model.

This test runs a very short chain very briefly, so it cannot test for inference correctness or convergence right now. This is only a test to execute all of the code and ensure that there are no errors or `NaN`s.

Differential Revision: D19144401

fbshipit-source-id: 9d5b2f53dcb6358c743cb790ebf3938b4147dab9