Add argument observe_model for HMM
Summary:
For the HMM, the model consists of `theta,mus, sigmas`.
`theta` is the transition matrix of the hidden state.
`mus` and `sigmas` parameterize the emission distributions for the `K` different hidden states.

This new model argument, `observe_model`, enables the HMM to perform two different inference problems. In one problem (`observe_model=True`) the model value is known to the inference algorithm, otherwise the model isn't known.

Currently, the default is `observe_model=True` because this is the problem of interests for demonstrating the usefulness of single-site MCMC.

Reviewed By: nimar

Differential Revision: D19314512

fbshipit-source-id: e50e43e17eb64eca17dc92419423c13ea572cd9d