HMC proposer (#92)
Summary:
Pull Request resolved: https://github.com/facebookincubator/BeanMachine/pull/92

The propose method uses the world functions to compute the potential energy. It then takes leap-frog steps, and returns the difference in kinetic energy. The potential energy will be calculated by the world during the accept / reject.

If a child's parents are changed during a step, an error will be raised.

Reviewed By: nazanint

Differential Revision: D18898569

fbshipit-source-id: b7f86c9f28d206444fdc575fdadbc046d1176efe