First attempt at rules engine
Summary:
I've created the outline of a simple, combinator-based rules engine for rewriting syntax trees.

The basic idea is:

* A *rule* is a *partial function*; typically our rules will be AST to AST projections. A rule is "partial" because a rule can "fail" or "succeed", and produces no transformed value if it fails.
* Rules can be implemented with *patterns*, from the previous diff. If the pattern matches the input then the rule runs a projection on the input, and succeeds. If the pattern does not match the input, the rule fails.
* Rules can be built out of other rules using *combinators*. A combinator is a function that takes some number of rules and produces a rule.  For example "try to apply this rule once; if it applies, use the result; if it fails, use the original value" is a combinator; it takes a rule, and it returns a new rule that has a more complex action.

If you're following along you've probably noticed that a "pattern" is logically just a rule such that if the pattern does not match, the rule fails, and if the pattern does match, the rule succeeds with an identity projection. Though this is certainly true, I would like to be able to develop, test and debug patterns separately from rules, so I'm not going to combine them.

Similarly you might notice that the transformations in the pattern library that simplify patterns could be implemented as rules! Though I might go there at some point, again, for simplicity I'm going to keep them separate at this point.

Reviewed By: nimar

Differential Revision: D17889634

fbshipit-source-id: cba48f620d12a5168a02f986f6a6ef34741f2557