Start creating a rewriting rules engine
Summary:
I've been thinking about how to structure a compiler that is mostly concerned with pattern recognition and Python-to-Python rewriting. The approach I'm going to try is a strategy-based rules engine, like Catalyst or Strageto. Though there is some up-front cost to developing a pattern matching rules engine, you get wins when your tree rewriting code is (1) concise, and (2) has the smarts built into the engine code, not the rules code.

I'm starting here with a pattern matching library. I wish to concisely represent structural patterns in trees, much as regular expressions concisely represent structural patterns in text. For example, suppose we have the problem "is this AST node of the form "0 + anything?"  For example, suppose we have this node in hand:

    node: Ast = BinOp(op=Add(), left=Num(n=0), right=Num(n=1))

I would like to be able to define a *pattern object* where the code that defines the pattern object is structurally similar to the code which defines the object being matched:

    b : Pattern = binop(op=Add, left=num(n=0), right=Any)

Notice the small differences -- the object *constructs* an `Add` whereas the pattern *uses the type*, for instance.

We can then say

    result = b.match(node)
    if result.is_success(): ...

I want the result to be not just True or False, but rather an object that describes the match and any submatches for complex patterns.

Also for debugging purposes I'd like to be able to dump patterns as text; I'll add graphs and trees later, but right now patterns dump themselves as (roughly!) the Python code they represent:

    print(str(b))

    (isinstance(test, BinOp) and isinstance(test.op, Add) and
    (isinstance(test.left, Num) and test.left.n==0))

Coming up next, I'll define a *rule* as a partial function from values to values that is constructed from a pattern and a projection. We'll then be able to express rules like "if the left side of the addition is zero, rewrite the expression to use only the right side".  That is, (0 + x) --> x, simplifying the tree.

Reviewed By: nimar

Differential Revision: D17870462

fbshipit-source-id: 4795b14f1ede6287ce725773c88e80c0cd05d243