Make rule combinators explicit in code
Summary:
In a combinator-based rules system, almost everything can be constructed out of a small handful of basic combinators; you need identity, fail, left choice, the y combinator and maybe a couple more, and from that you can build *everything*. That has an attractive simplicity, but it also makes the code harder to understand and debug; some rule isn't doing what you think it should and you look at it and see nothing but a hundred compositions of left-choice, and the semantics are gone.

It's arguably better to have a large number of explicitly implemented combinators; they can be self-describing, they have source code locations for breakpoints, and so on.

I've converted the combinators that *could* be implemented as invocations of other combinators into classes of their own, and added a test that shows the structure of a rule as a string.

Reviewed By: nimar

Differential Revision: D17918491

fbshipit-source-id: 8c001ca211dcd7a00c425944946835f970172660