General adaptive inference for Bean Machine
Summary:
This diff enables specific inference methods to implement adaptation. Specifically this includes

- Add accept/reject results to `MonteCarloSamples`
- Add information about the adaptation period to `MonteCarloSamples`, and use this information to exclude warm-up samples by default.
- Define the API for `do_adaption()`, a proposer-specific method, and call this method within the single-site inference loop.

This does not change the existing API for `MonteCarloSamples` as all.

This diff is also a general way to add additional information to `MonteCarloSamples`. For example, the accept/reject probabilities could be stored in the future, in the same way that accept/reject results are stored in this diff.

Reviewed By: nazanint, yucenli

Differential Revision: D19602648

fbshipit-source-id: fa9cd4790a9ee19f2363f417c5c93cfd32c88bae