Implement "all children" combinator to work on lists
Summary:
The "all children" combinator that takes a rule and applies it to every child of a node now works when given a list.

However, it does not yet handle the case where the rewrite rule wishes to insert new elements or delete old elements from a list.  That will come in a subsequent diff.

UPDATE: Upon reflection, this is an awkward way to implement this combinator, but fixing it at this point in the stack would require some unnecessarily difficult merge conflicts.

The better way to implement this is to have two combinators, one that handles exclusively members of lists, and one that handles exclusively children of terms.  The "all_children" combinator is then just the composition `all_children(r) = if_then(is_list, AllListMembers(r), AllChildren(r))`.

I will make that change in a future diff in this stack.

Reviewed By: nimar

Differential Revision: D17939158

fbshipit-source-id: 3d2718bd35bcb7c7f7a35a9e9276ac81f8d94e8d