Implement the SomeChildren combinator
Summary:
The SomeChildren combinator takes a rule and succeeds if the rule can be successfully applied to at least one child.

This combinator is useful for attaining fixpoints.

UPDATE: As noted in the previous diff, this is an awkward implementation that I will fix in a future diff rather than taking the merge conflict now.

Reviewed By: nimar

Differential Revision: D17955907

fbshipit-source-id: b31d3039fdde3676915fa0ac25038e9c3c69f4f5