Implement rule combinator for applying rules to children in a tree
Summary:
This is in a sense the most important rule combinator because (1) it is fundamental; there's no
easy way to reduce it to a simpler rule combinator, and (2) it is what allows us to rewrite throughout
a tree, not just at the root.

The combinator is conceptually straightforward. Given a rule it produces a new rule which is "apply the rule to all children of the test; whatever results those rule applications produce become the new children".

Since this involves both destructuring a node to get its children, and constructing a new one, and there may be different ways of doing that for different domains, I've added a new concept, the "rule domain"; a rule domain is just a pair of functions that describe how to destructure a node and construct a new node of a given type with given children.  I've created an implementation for AST nodes.

We'll need to extend this to support lists; the children of a list are of course its elements. Lists are interesting because we'll want to be able to express not just replacement but deletion and insertion. I'll do that in a later diff.

Reviewed By: nimar

Differential Revision: D17936760

fbshipit-source-id: 6d8d49584733ed5dd6b02ec082f1c0eaaba80e8f