Add ability to add or delete items from lists with special rules and combinators
Summary:
A child rule of the AllChildren, SomeChildren, or OneChild combinators may return as its result a ListEdit object which contains either an empty list -- indicating that the rule wishes the current item to be deleted from the list -- or multiple items, indicating that the rule wishes multiple items to be spliced in.

This is convenient but a bit dodgy, as you can see from the test case we get strange results when the outer rule is not an application of the AllChildren, SomeChildren or OneChild combinator.

If this becomes an issue, I will probably add traditional functional-style "head : tail" pattern matching to the pattern library as well; that has better behaviour under composition.

Differential Revision: D17943059

fbshipit-source-id: 935108ed2a304928c570ca4baf41ae72fb79cd65