Implement simple infinite-loop detection for TryMany combinator
Summary:
One of the benefits of a combinator-based approach is that you can do analysis while building combinators to look for inefficiencies or errors.

While building a test case for the previous diff I accidentally created a rule that went into an infinite loop; we now detect that situation and raise an error when the rule is constructed.

Reviewed By: michaeltingley

Differential Revision: D17961514

fbshipit-source-id: 1bb59b8f1ded358aaaf132204c2e87f713fed02c