Implement predicate matching rules and patterns
Summary:
I've implemented a predicate and projection rules so that we can construct rules like "for all the even numbers in the program, add one".

A rule is typically a combination of a predicate formed from a pattern, and a projection; this just allows us to split those two things apart. This is typically not necessary, but there may be cases where it is more efficient, or more concisely expressed, to do so.

Reviewed By: michaeltingley

Differential Revision: D17963472

fbshipit-source-id: 763aac7c9bacca7d3e9f603630caa6a2949600f8