Add more support to AST pattern library
Summary:
I've added a few more nodes to my AST pattern library; eventually we should probably just auto-generate this, as it is straightforward boilerplate.  (In fact, doing so could be a test of the rules engine!)

I've also added a test that demonstrates a concise rewrite: find all the functions in a module with a decorator; delete the ones that lack the decorator and replace the bodies of those with the decorator with "pass".

Reviewed By: michaeltingley

Differential Revision: D17968036

fbshipit-source-id: b2a21d408c662b68f9d3f18db8cb65624ce97ec7