Start building a constant folder
Summary:
When analyzing a BeanMachine program it will be helpful if we can reduce constant expressions; this will make the graph smaller and will save us work in writing a graph optimizer.

Here I've just got a very straightforward optimizer that only works on numbers and bools, and requires that every node in the expression be a constant expression.

Later on I'll show how to build a rewriter that can handle situations like simplifying `1 + x + 2` to `3 + x`, simplifying `1 + (x - 1*1)` to `x`, and so on.

I haven't implemented folding on comparison operators yet because they are weird in Python; it will take some more gear to get that right.

Reviewed By: michaeltingley

Differential Revision: D17996921

fbshipit-source-id: 5ed6e9addbda3221a0f633d05b37190e60b3993b