Refactor all/some/one child combinators into multiple classes.
Summary:
As I noted a few diffs back in this stack, I realized recently that putting the "all/some/one children of a term" and the "all/some/one members of a list" semantics into the same combinators was a bad idea; these are logically separate, and can be combined with a simple combinator when we need to.

This refactors them into several classes that each does one thing.

Reviewed By: michaeltingley

Differential Revision: D18019958

fbshipit-source-id: f1005389dac70e697e2df09e577a6521dd014c3e