Start writing single-assignment transformation rules
Summary:
It will be helpful when transforming from BM to BMG to have the program in a simple version of "single assignment" form.  That is, if we have a program something like:

    x = a() + b()
    return x + 2

then it will be convenient to have this in the equivalent form

    t1 = a()
    t2 = b()
    x = t1 + t2
    t3 = 2
    t4 = x + t3
    return t4

I've started a proof-of-concept here by turning return statements into this form.

Reviewed By: michaeltingley

Differential Revision: D18229330

fbshipit-source-id: ca418854018544426ca08ada07741bb5276ef855