Implement combinator that enables treating members of lists as children
Summary:
We have some AST nodes that have both "normal" term children and list children. For example, comparison is n-ary in Python and has three children, `ops`, `left` and `comparators`.  `ops` and `comparators` are lists, `left` is a normal child.

This makes it awkward to write rules that apply to all children of a thing; the usual `all` combinator says that a `Compare` has exactly three children, but if we are rewriting children of a comparison it is convenient to be able to rewrite the *logical* children rather than the actual children.

I've added a combinator that facilitates this scenario; in an upcoming diff I'll use it to write an interesting recursive constant folder for comparison operations.

Reviewed By: michaeltingley

Differential Revision: D18021319

fbshipit-source-id: 4f5e19362819e796a55b18390d0062c8c73258be