Add some_bottom_up combinator
Summary: The some_bottom_up combinator is the leaves-to-root variation on the some_top_down combinator. It applies a rule to every node in the tree, leaves to root, ignores nodes to which it does not apply, and fails iff no rule application succeeded. It makes as much progress as possible and fails if none could be made, so you can construct a fixpoint combinator out of it.

Reviewed By: michaeltingley

Differential Revision: D18094951

fbshipit-source-id: 2e2a9084f68bf94dd5e208c7646715314728d60f