Add at_least_once rule combinator
Summary: The at_least_once combinator does what it says on the tin: applies a rule, fails if the rule fails, and if it succeeds, keeps trying it until it succeeds.  This combinator is useful when making fixpoints for rules that can keep doing work, but we wnat to know if no work was done.

Reviewed By: michaeltingley

Differential Revision: D18092269

fbshipit-source-id: 90b35ed4474d1c54f4395ce48622603ae56a2dd3