Create SomeOf rule combinator
Summary: The SomeOf rule combinator is an extension of either_or_both; it takes a list of rules and tries to apply each in order, with the output of earlier rules becoming the input of later rules. The new rule succeeds if at least one of the given rules succeeds, otherwise it fails.

Reviewed By: michaeltingley

Differential Revision: D18214152

fbshipit-source-id: a7b3a8c4f2d683111432ec3ba23a599bea0ef66c