Add constant folding for comparison expressions
Summary:
Python comparison expressions are unusual in that they are n-ary expressions:

    x < y < z

has the semantics

    x < y and y < z

This means that we have a great example of using a list pattern to first transform the operator into the binary form, and then optimize it once it is in the binary form.

However, we'll also see that doing so causes our constant folder to stop reaching a fixpoint. We'll take care of that in a subsequent diff.

Reviewed By: michaeltingley

Differential Revision: D18046766

fbshipit-source-id: 8833cb09db49fa5fe93c1259d5bba72294651d33