Add some_top_down combinator
Summary:
The top_down combinator applies a rule to every node in the tree, root to leaves, and requires that it succeed on every node. This means that typically you apply top_down to a rule that cannot fail. But this then presents a problem: if the rule cannot fail then top_down(rule) cannot fail, which means that we cannot know if the rule did anything. If this does not reach a fixpoint we have no way to iterate until it does.

some_top_down solves this problem. It applies a rule to every node in the tree, root to leaves, ignores nodes for which it does not succeed, and the whole thing succeeds if and only if the rule was successfully applied to at least one node. That is, it does the maximum amount of work possible, and fails only if no work was done.  That way we can build a fixpoint combinator: TryMany(some_top_down(rule)) is a fixpoint combinator.

We'll add the equivalent bottom-up combinator in a subsequent diff.

Reviewed By: michaeltingley

Differential Revision: D18094263

fbshipit-source-id: d2ef09c236aef54d2e4b4f746f6f18c72a663c63