Major refactoring of constant folder
Summary:
This is a major refactoring of the constant folder; as is often the case in refactorings, we can both add new features and eliminate a bunch of specific code by solving a more general problem. In this case, we take advantage of Python's loose attitude towards typing to extend constant folding from just special cases of bools, ints, floats and tensors, to all operations involving any combinations of those four types. We delete a bunch of special-purpose code along the way.

The new strategy is simply to identify unary, binary and n-ary operators that have all constant operands, convert the constant operands in source code to actual constants in memory, do the operations, and then convert the results back to source code.

An interesting point is that since some constant foldings can fail -- because the program is wrong -- this means that the constant folding can also fail at fold time. We need to ensure that our folder is robust against these scenarios in two ways: first, that the folder does not itself throw, and second, that it never gets stuck in a loop of trying to fold, failing, and then trying again.

Reviewed By: michaeltingley

Differential Revision: D18260163

fbshipit-source-id: adbafde5934a39dbbd2a358a7b6c2f945084d381