Create a pattern for more constant tensors
Summary:
I want to do more sophisticated folding of constant tensors, which means that we'll need to be able to recognize them. I've added a recursive pattern for recognizing lists that contain only constant numbers or lists of constant numbers, which will be a good first approximation of what is a legal constant tensor.

I'll fix up the rules to do folding on constant tensors in the upcoming diffs.

Reviewed By: michaeltingley

Differential Revision: D18235972

fbshipit-source-id: 4dac12af052c61fa7a36c9f7578282be77dc16df