Add new combinator for turning exceptions into failures
Summary: Sometimes a pattern rule will match the pattern but then the projection will fail with an exception through no fault of our own; it can be expensive or impossible to know if a particular operation is going to fail.  In those cases we can treat the rule as failing.

Reviewed By: michaeltingley

Differential Revision: D18275964

fbshipit-source-id: da05167fa653af2baae1698163990359ee79c879