Single-assignment rewriter should skip constant tensors
Summary: My simplified single assignment form rewriter seeks to get all assignments into a form where the right side of the assignment is a single operation where the operands are all identifiers. However, we should skip this rewrite if the right side of an assignment is a constant tensor. Constant tensors look like function calls, but logically they are atomic values and we can keep them as such.

Reviewed By: michaeltingley

Differential Revision: D18355828

fbshipit-source-id: 14113c9b380e27dadc758d210cb8475e44e59415