First steps towards BM to BMG compiler
Summary:
I've started a proper BM to BMG compiler; in this diff, I reduce a BM program to my simplified single assignment form and ensure that (1) all subtractions are realized as additions with a negation on the right side, and (2) all negative constants are *not* turned into constants with negation operators on them.

Our goal here is to get the program into a form where there is at most one operation per statement, and that operation is one of the operations supported by BMG (addition, multiplication, exp, negation).

In an upcoming diff I will then transform each line into a constructor of a node in the graph; when we run the resulting program, we will construct the graph.

I have not yet added an error detection pass which will tell the user if they've used an operation not (yet) supported in BMG; that will come in a (much) later diff.

Reviewed By: michaeltingley

Differential Revision: D18354931

fbshipit-source-id: 095a365306d0447fd49f60a88dfd0e794dd11671