Add the descend_until combinator
Summary:
The descend_until(test, rule) combinator descends through the tree, running the test rule on each node. If the test succeeds, then we stop descending and run the given rule on the node which passed the test. We do this to all the nodes in the tree that match the test, starting from the root.

This combinator is useful for tasks like "find all the functions that match a predicate in this module and rewrite their bodies, but do not apply the rewriting to code outside of functions", for instance.

Reviewed By: michaeltingley

Differential Revision: D18364771

fbshipit-source-id: fd3a2d0586b409776768f17ed496e55ea53d1306