Add specific_child rule combinator
Summary: The `specific_child` rule combinator is convenient for those situations where you have determined that a predicate is true of a parent, and now you want to apply a rewriting rule to a specific child. For instance, maybe we determine that a function has a particular decorator, and then we wish to apply a rewrite, but just to the decorator list; it can be awkward to describe such a rewrite using the existing pattern rule mechanism.

Reviewed By: michaeltingley

Differential Revision: D18404607

fbshipit-source-id: 482b62fca9f624fdbd831bcd9f7e9e3a85d66703