Single assignment form supports function calls
Summary:
My simplified single assignment form rewriter now deals with argument lists of function calls.  It does not yet deal with receivers of function calls.

That is, if we have something like `y = foo(1 + x)` this will be rewritten to

    t1 = 1
    t2 = t1 + x
    y = foo(t2)

which preserves the desired invariant that every line is an assignment where the right side of the assignment only involves a single operation whose operands are all identifiers.

Reviewed By: michaeltingley

Differential Revision: D18351255

fbshipit-source-id: 77d87e3be37ff014b42157998ec61c0d07d79a5f