Graph builder now folds constant operations as it goes
Summary:
As I noted in the previous diff, the Beanstalk constant folder does not do constant propagation across function calls, which means that we can end up in a situation where the graph produced has unfolded constant operations, unnecessary TO_REAL nodes, and so on.  To fix this, I've added constant folding to the graph builder, and it now produces much simpler graphs. However, a side effect of this is that nodes which were added to the graph and then "orphaned" by optimizations stick around. I'll add a pass to remove them in a subsequent diff.

For example, this graph that was produced in the previous diff:

{F222219706}

is now greatly simplified, but leaves behind a bunch of orphan nodes:

{F222228360}

Reviewed By: michaeltingley

Differential Revision: D18465426

fbshipit-source-id: 8cae5afb96f68870cd2f264f5efdfa0cf5a851d6