Minor change to beanstalk code generation
Summary: The BeanMachine to BMGraph compiler needed a small change in its code generation; when we execute the functions that fill in the graph builder, I'll need to know what those "top level" nodes were so that I can strip out the orphans generated in the previous diff.  We now put those nodes in a list of roots.

Reviewed By: michaeltingley

Differential Revision: D18483362

fbshipit-source-id: 11534020f934735f055295b5a362d656abf1a313