Add test case showing Beanstalk's correct treatment of parameterized samples
Summary:
When the sample function takes a parameter, as in the test case added here, we need to ensure that we are adding as many SAMPLE nodes to the graph as there are samples created in the program; fortunately, since we now memoize hoisted constant nodes, we get this working "for free".

BMG does not yet support random variables that are parameterized on *samples*, but when we do, it will hopefully require only minimal changes to the compiler; we should be able to pass in a sample node just as easily as a constant node here.

The test case is:

    sample
    def x(n):
      return Bernoulli(tensor(0.5) + n * tensor(0.1))

    sample
    def z():
      return Bernoulli(x(0) * tensor(0.3) + x(0) * tensor(0.1) + x(1) * tensor(0.4))

The graph generated by the test case looks like this:

{F222219706}

Notice that though it is fully deduplicated, it is NOT constant folded because the Beanstalk constant folder does not fold across function call boundaries. (And nor should it.)  However, we can simplify this graph further, which we can do in the graph builder itself; I'll add those features in upcoming diffs.

Reviewed By: michaeltingley

Differential Revision: D18462622

fbshipit-source-id: e58582bedab983f840c56f0f9704aa0fb184e792