Deduplicate nodes in Bean Machine Graph representation
Summary:
My builder now deduplicates non-sample nodes when constructing a BMGraph object, resulting in a simpler graph. Doing so is straightforward because we always build the DAG from the leaves up, so simply memoizing the construction functions is sufficient.

Note that we never deduplicate sample nodes; samples are by definition distinct samples from a distribution. Notice in the example code that even though we have two samples drawn from (what becomes) the same Bernoulli, we still have two samples.  And therefore all the operations that consume one of those samples are also not deduplicated.

That is, we go from this tree:

{F222188875}

To this deduplicated DAG:

{F222189114}

Reviewed By: michaeltingley

Differential Revision: D18456938

fbshipit-source-id: 13a819d6657c2c27bc59274ccc1a23b9866c2f31